RGBオブジェクトを作成 FizzBuzztestのファイルを分割する
