+++ lib/fizz_buzz.rb
@@ -32,23 +32,3 @@ def fizz_buzz(n)
     n.to_s
   end
 end
-
-# テストを記載する
-# テストの簡単な手順
-# 1. Minitestなどのテスティングフレームワークを呼ぶ
-# 2. classをキャスパルケースで定義する
-# 3. クラス内にテスティングメソッドを定義する
-
-require 'minitest/autorun'
-
-class FizzBuzzTest < Minitest::Test
-  def test_fizz_buzz
-    assert_equal '1', fizz_buzz(1)
-    assert_equal '2', fizz_buzz(2)
-    assert_equal 'Fizz', fizz_buzz(3)
-    assert_equal '4', fizz_buzz(4)
-    assert_equal 'Buzz', fizz_buzz(5)
-    assert_equal 'Fizz', fizz_buzz(6)
-    assert_equal 'Fizz Buzz', fizz_buzz(15)
-  end
-end
