+++ lib/rgb.rb
@@ -1,0 +1,28 @@
+# RGBカラー変換プロプログラム
+# - Red
+#     - R=65 => 10進数の整数
+# - Green
+#     - G=105 => 10進数の整数
+# - Blue
+#     - B=225 => 10進数の整数
+# - システム要件
+#     - 10進数を16進数に変換するto_hexメソッド
+#         - to_hexメソッドは3つの整数を受け取る。
+#         - それぞれを16進数に変換した文字列を返し、文字列の先頭には“#”を付ける。
+#     - 16進数を10進数に変換するto_intsメソッド
+#         - to_intsメソッドはRGBカラーを表す16進数の文字列を受け取る。
+#         - R、G、Bのそれぞれを10進数の整数に変換した値を配列として返す。
+#         ´´´Ruby
+#         to_hex(0, 0, 0) #=> "#000000"
+#         to_hex(255, 255, 255) #=> "#ffffff"
+#         to_hex(4, 60, 120) #=> "#043c78"
+#         to_ints('#000000') #=> [0, 0, 0]
+#         to_ints('#ffffff') #=> [255, 255, 255]
+#         to_ints('#043c78') #=> [4, 60, 120]
+
+def to_hex(r, g, b)
+  '#' +
+  r.to_s(16).rjust(2, '0') +
+  g.to_s(16).rjust(2, '0') +
+  b.to_s(16).rjust(2, '0')
+ end
